requirements.txt
django==3.2.23
djangorestframework==3.12.4
pandas==1.5.3
numpy==1.23.5
geopy==2.3.0
requests==2.28.2
gunicorn==20.1.0
python-decouple